(Future Fun) TODO: come back and find/fix the bug
There are at least two bugs. One is simple, it causes an infinite
loop.

The other is akward/invisible. It kills my debugger, randomly,
either on or just after line 61. I spent a little while trying to
figure it, but now it's just a waste of time. So i'll leave it to
my future self to come back, find, and fix it.
At current, I am going to re-do the whole challenge, possibly with
some different logic. Either way, if it was some random typo, then
it'll be resolved.